+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## Next version
 
+- Motorider map theme improvements
 - Rename `MapDatabase` to `MapFile` [#1184](https://github.com/mapsforge/vtm/pull/1184)
   - Rename `MultiMapDatabase` to `MultiMapFile`
 - Minor improvements and bug fixes