+++ docs/Changelog.md
@@ -2,7 +2,7 @@
 
 ## Next version
 
-- Motorider map theme improvements [#1185](https://github.com/mapsforge/vtm/pull/1185)
+- Motorider map theme improvements [#1183](https://github.com/mapsforge/vtm/issues/1183)
 - Rename `MapDatabase` to `MapFile` [#1184](https://github.com/mapsforge/vtm/pull/1184)
   - Rename `MultiMapDatabase` to `MultiMapFile`
 - Minor improvements and bug fixes