+++ docs/Changelog.md
@@ -5,6 +5,7 @@
 - Hillshading from HGT digital elevation model data [#1189](https://github.com/mapsforge/vtm/pull/1189)
   - `vtm-hillshading` module
 - Motorider map theme improvements [#1183](https://github.com/mapsforge/vtm/issues/1183)
+- Biker map theme improvements [#1192](https://github.com/mapsforge/vtm/issues/1192)
 - Rename `MapDatabase` to `MapFile` [#1184](https://github.com/mapsforge/vtm/pull/1184)
   - Rename `MultiMapDatabase` to `MultiMapFile`
 - Minor improvements and bug fixes